requirements.txt
python-nmap>=0.7.1
netifaces>=0.11.0
pymetasploit3>=1.0.3
pyshark>=0.5.3
scapy>=2.5.0
